Fixing #74 event logs still collecting after unchecking shutdown option
Investigated this issue. Yes, it is a byproduct of the same issue as #54 . The top , group-level setting does not really impact the enabled/disabled. It has to be done individually. Therefore, I will change it for this particular case, but the large issue for all groups has be considered and resolved .